+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * [Fix #379] KIS functionality interferes with the stock construction mode.
 * [Fix #380] Model equippable parts cannot be carried.
 * [Fix #381] Hide UI command doesn't play well with KIS.
+* [Fix #382] Static attach doesn't verify which joint is broken.
 
 # 1.27 (December 23rd, 2020):
 * [Enhancement] Check for colliders when node attaching a part.