+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.28 (pre-release):
+# 1.28 (January 15th, 2021):
 * [Fix #378] Dropped parts cannot be moved or trigger physics.
 * [Fix #379] KIS functionality interferes with the stock construction mode.
 * [Fix #380] Model equippable parts cannot be carried.